% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_library.R
\name{reorder_library}
\alias{reorder_library}
\title{Réorganise les appels à \code{library}.}
\usage{
reorder_library()
}
\description{
Réorganise les appels à \code{library}.
}
\examples{
# Utilisez le menu "Addins" dans
# RStudio pour exécuter la fonction !
}
