% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_addin.R
\name{geocode_addin}
\alias{geocode_addin}
\title{Géocode un point placé sur une carte}
\usage{
geocode_addin()
}
\description{
Géocode un point placé sur une carte
}
\examples{
\dontrun{

if (interactive()) {

geocode_addin()

}

}
}
