% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_library.R
\name{reorder_library}
\alias{reorder_library}
\title{Réorganise les appels à \code{library}.}
\usage{
reorder_library()
}
\description{
Place les appels à \code{library} disséminés dans un
 script au début et les remplace par une section contenant le nom du package.
}
\examples{
# Utilisez le menu "Addins" dans
# RStudio pour exécuter la fonction !
}
